{"schema":"decodex/commit/1","summary":"Modularize frozen frame pixel-buffer bridge","authority":"manual"}
