+++ native/macos-host/Sources/RsnapNativeHostKit/FrozenFrameAuthority.swift
@@ -97,36 +97,6 @@ final class FrozenFrameAuthority: @unchecked Sendable {
 		let startedAtUptime: TimeInterval
 	}
 
-	private final class PixelBufferImageBacking {
-		let pixelBuffer: CVPixelBuffer
-		let baseAddress: UnsafeMutableRawPointer
-		let byteCount: Int
-		let unlockFlags = CVPixelBufferLockFlags.readOnly
-
-		init?(_ pixelBuffer: CVPixelBuffer) {
-			guard CVPixelBufferLockBaseAddress(pixelBuffer, unlockFlags) == kCVReturnSuccess else {
-				return nil
-			}
-			guard let baseAddress = CVPixelBufferGetBaseAddress(pixelBuffer) else {
-				CVPixelBufferUnlockBaseAddress(pixelBuffer, unlockFlags)
-				return nil
-			}
-			let height = CVPixelBufferGetHeight(pixelBuffer)
-			let bytesPerRow = CVPixelBufferGetBytesPerRow(pixelBuffer)
-			guard height > 0, bytesPerRow > 0 else {
-				CVPixelBufferUnlockBaseAddress(pixelBuffer, unlockFlags)
-				return nil
-			}
-			self.pixelBuffer = pixelBuffer
-			self.baseAddress = baseAddress
-			self.byteCount = bytesPerRow * height
-		}
-
-		deinit {
-			CVPixelBufferUnlockBaseAddress(pixelBuffer, unlockFlags)
-		}
-	}
-
 	private struct PreparedContentFilter {
 		let filter: SCContentFilter
 		let selfCaptureFilterComplete: Bool
@@ -854,7 +824,7 @@ final class FrozenFrameAuthority: @unchecked Sendable {
 			return nil
 		}
 		guard
-			let rgb = Self.rgbSample(
+			let rgb = FrozenFramePixelBufferBridge.rgbSample(
 				from: record.pixelBuffer,
 				point: point,
 				displayFrame: record.displayFrame
@@ -878,7 +848,7 @@ final class FrozenFrameAuthority: @unchecked Sendable {
 		guard let record else {
 			return nil
 		}
-		return Self.loupePatch(
+		return FrozenFramePixelBufferBridge.loupePatch(
 			from: record.pixelBuffer,
 			point: point,
 			displayFrame: record.displayFrame,
@@ -926,7 +896,9 @@ final class FrozenFrameAuthority: @unchecked Sendable {
 		}
 		stateLock.unlock()
 
-		guard let record, let image = Self.makeImage(from: record.pixelBuffer) else {
+		guard let record,
+			let image = FrozenFramePixelBufferBridge.makeImage(from: record.pixelBuffer)
+		else {
 			return nil
 		}
 		return FrozenFrameSnapshot(
@@ -1401,123 +1373,6 @@ final class FrozenFrameAuthority: @unchecked Sendable {
 		)
 	}
 
-	private static func makeImage(from pixelBuffer: CVPixelBuffer) -> CGImage? {
-		guard let backing = PixelBufferImageBacking(pixelBuffer) else {
-			return nil
-		}
-		let width = CVPixelBufferGetWidth(pixelBuffer)
-		let height = CVPixelBufferGetHeight(pixelBuffer)
-		let bytesPerRow = CVPixelBufferGetBytesPerRow(pixelBuffer)
-		guard width > 0, height > 0, bytesPerRow >= width * 4 else {
-			return nil
-		}
-		let retainedBacking = Unmanaged.passRetained(backing)
-		guard
-			let provider = CGDataProvider(
-				dataInfo: retainedBacking.toOpaque(),
-				data: backing.baseAddress,
-				size: backing.byteCount,
-				releaseData: { info, _, _ in
-					guard let info else {
-						return
-					}
-					Unmanaged<PixelBufferImageBacking>.fromOpaque(info).release()
-				}
-			)
-		else {
-			retainedBacking.release()
-			return nil
-		}
-		let bitmapInfo = CGBitmapInfo.byteOrder32Little
-			.union(CGBitmapInfo(rawValue: CGImageAlphaInfo.premultipliedFirst.rawValue))
-		return CGImage(
-			width: width,
-			height: height,
-			bitsPerComponent: 8,
-			bitsPerPixel: 32,
-			bytesPerRow: bytesPerRow,
-			space: CGColorSpace(name: CGColorSpace.sRGB) ?? CGColorSpaceCreateDeviceRGB(),
-			bitmapInfo: bitmapInfo,
-			provider: provider,
-			decode: nil,
-			shouldInterpolate: false,
-			intent: .defaultIntent
-		)
-	}
-
-	private static func rgbSample(
-		from pixelBuffer: CVPixelBuffer,
-		point: CGPoint,
-		displayFrame: CGRect
-	) -> RGBSample? {
-		let width = CVPixelBufferGetWidth(pixelBuffer)
-		let height = CVPixelBufferGetHeight(pixelBuffer)
-		let bytesPerRow = CVPixelBufferGetBytesPerRow(pixelBuffer)
-		guard width > 0, height > 0, bytesPerRow >= width * 4 else {
-			return nil
-		}
-		guard CVPixelBufferLockBaseAddress(pixelBuffer, .readOnly) == kCVReturnSuccess else {
-			return nil
-		}
-		defer {
-			CVPixelBufferUnlockBaseAddress(pixelBuffer, .readOnly)
-		}
-		guard let baseAddress = CVPixelBufferGetBaseAddress(pixelBuffer) else {
-			return nil
-		}
-		return try? RsnapBgraFrameSampler.rgbSample(
-			width: width,
-			height: height,
-			bytesPerRow: bytesPerRow,
-			baseAddress: baseAddress,
-			byteCount: bytesPerRow * height,
-			displayFrame: displayFrame,
-			point: point
-		)
-	}
-
-	private static func loupePatch(
-		from pixelBuffer: CVPixelBuffer,
-		point: CGPoint,
-		displayFrame: CGRect,
-		sidePixels: Int
-	) -> CGImage? {
-		let width = CVPixelBufferGetWidth(pixelBuffer)
-		let height = CVPixelBufferGetHeight(pixelBuffer)
-		let bytesPerRow = CVPixelBufferGetBytesPerRow(pixelBuffer)
-		let side = max(sidePixels, 1)
-		guard width > 0, height > 0, bytesPerRow >= width * 4 else {
-			return nil
-		}
-		guard CVPixelBufferLockBaseAddress(pixelBuffer, .readOnly) == kCVReturnSuccess else {
-			return nil
-		}
-		defer {
-			CVPixelBufferUnlockBaseAddress(pixelBuffer, .readOnly)
-		}
-		guard let baseAddress = CVPixelBufferGetBaseAddress(pixelBuffer) else {
-			return nil
-		}
-		guard
-			let patch = try? RsnapBgraFrameSampler.loupePatch(
-				width: width,
-				height: height,
-				bytesPerRow: bytesPerRow,
-				baseAddress: baseAddress,
-				byteCount: bytesPerRow * height,
-				displayFrame: displayFrame,
-				point: point,
-				sidePixels: side
-			)
-		else {
-			return nil
-		}
-		return NativeHostImageBridge.cgImage(
-			width: patch.width,
-			height: patch.height,
-			rgba: patch.rgba
-		)
-	}
 }
 
 private final class FrozenFrameStreamOutput: NSObject, SCStreamOutput, SCStreamDelegate,
